// packages/node-sdk/src/rpc-types.ts
export interface SessionIdRpcInput {
  readonly sessionId: string;
}
// ...
export interface ReconnectMcpServerRpcInput extends SessionIdRpcInput {
  readonly name: string;
}